+++ axum-extra/CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning].
 - **breaking:** `CachedRejection` has been removed ([#699])
 - **breaking:** `<Cached<T> as FromRequest>::Rejection` is now `T::Rejection`. ([#699])
 - **breaking:** `middleware::from_fn` has been moved into the main axum crate ([#719])
+- **breaking:** `HasRoutes` has been removed. `Router::merge` now accepts `Into<Router>` ([#819])
+- **breaking:** `RouterExt::with` method has been removed. Use `Router::merge` instead. It works
+  identically ([#819])
 
 [#666]: https://github.com/tokio-rs/axum/pull/666
 [#699]: https://github.com/tokio-rs/axum/pull/699
@@ -21,6 +24,7 @@ and this project adheres to [Semantic Versioning].
 [#756]: https://github.com/tokio-rs/axum/pull/756
 [#790]: https://github.com/tokio-rs/axum/pull/790
 [#816]: https://github.com/tokio-rs/axum/pull/816
+[#819]: https://github.com/tokio-rs/axum/pull/819
 
 # 0.1.2 (13. January, 2022)